[WNMGDS-971] Convert Drawer and HelpDrawer to Typescript (#1466)
* Convert Drawer to TS.

* Refactor HelpDrawer to TS.

* Update Drawer to use correct ref

* Updated types in tests and realized toggles didn't have html el props

Was getting this error:
```
packages/design-system/src/components/Drawer/DrawerToggle.test.tsx:52:46 - error TS2345: Argument of type '{ ariaLabel: string; }' is not assignable to parameter of type 'Partial<DrawerToggleProps>'.
  Object literal may only specify known properties, and 'ariaLabel' does not exist in type 'Partial<DrawerToggleProps>'.

52     const { wrapper } = renderDrawerToggle({ ariaLabel });
                                                ~~~~~~~~~
```

* Export all the things

* Whoops, fix unit test

* Forgot that we replace this prop by one with a different name
