+++ packages/design-system-docs/src/pages/components/Drawer/_Drawer.docs.scss
@@ -17,7 +17,7 @@ Style guide: components.drawer
 
 @react-example DrawerDefault.example.jsx
 
-@react-props Drawer.jsx
+@react-props Drawer.tsx
 
 Style guide: components.drawer.react-drawer
 */
@@ -47,7 +47,7 @@ Style guide: components.drawer.react-drawer-focus-trap
 /*
 `<DrawerToggle>`
 
-@react-props DrawerToggle.jsx
+@react-props DrawerToggle.tsx
 
 Style guide: components.drawer.react-drawer-toggle
 */
@@ -59,7 +59,7 @@ This component is an enhanced `<Drawer>` component, providing the same propertie
 
 @react-example HelpDrawer.example.jsx
 
-@react-props HelpDrawer.jsx
+@react-props HelpDrawer.tsx
 
 Style guide: components.drawer.react-help-drawer
 */
