--- packages/design-system/src/components/Drawer/index.js
@@ -1,2 +1,0 @@
-export { default as Drawer } from './Drawer';
-export { default as DrawerToggle } from './DrawerToggle';
