+++ packages/design-system/src/components/Drawer/index.ts
@@ -1,0 +1,2 @@
+export * from './Drawer';
+export * from './DrawerToggle';
