--- packages/design-system/src/components/HelpDrawer/index.js
@@ -1,2 +1,0 @@
-export { default as HelpDrawer } from './HelpDrawer';
-export { default as HelpDrawerToggle } from './HelpDrawerToggle';
