+++ packages/design-system/src/components/HelpDrawer/index.ts
@@ -1,0 +1,2 @@
+export * from './HelpDrawer';
+export * from './HelpDrawerToggle';
