+++ packages/design-system/src/types/index.d.ts
@@ -1,8 +1,6 @@
 /**
  * index.d.ts - Typescript definition files entry point
  */
-export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
-export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
 
 // Type definitions generated by cmsds scripts
 export * from './Accordion';
@@ -13,10 +11,12 @@ export * from './Button';
 export * from './ChoiceList';
 export * from './DateField';
 export * from './Dialog';
+export * from './Drawer';
 export * from './Dropdown';
 export * from './FilterChip';
 export * from './FormControl';
 export * from './FormLabel';
+export * from './HelpDrawer';
 export * from './InlineError';
 export * from './MonthPicker';
 export * from './Pagination';
