+++ packages/ds-healthcare-gov/src/components/Header/DeConsumerMessage.tsx
@@ -3,7 +3,7 @@ import { TFunction, withTranslation } from 'react-i18next';
 
 interface DeConsumerMessageProps {
   deBrokerName: string;
-  /** 
+  /**
    * @hide-props
    */
   t: TFunction;
